Fixed code style bugs

+++ 0x13-more_singly_linked_lists/1-listint_len.c
@@ -8,13 +8,12 @@
  */
 size_t listint_len(const listint_t *h)
 {
-	size_t ele = 0;
-	
+	int ele = 0;
+
 	while (h)
 	{
 		h = h->next;
 		ele++;
 	}
-	
 	return (ele);
 }
